fix(agent-cli): reconcile the merged policy/spend-count tests and the r8 cap
updater_status has no renderer invoke() call site (its only consumer is
this agent surface's own updater_install confirm proof), so it needs the
same ALLOWLISTED_UNCATALOGUED entry boards_list/privacy_clear_data
already have; give notifications_mark_read/markAllRead the TSDoc the
catalogue coverage test expects for a described Irreversible row.

ai_spend_summary's grace-window proof count drops from 10 to 9: issue #1169
moved help_search off Irreversible, taking its charge_provider_daily proof
read with it. Regenerate the catalogue/API docs for the new TSDoc.

Splits mcp.rs's reply-shaping (tool_result/oversized_result/busy_result/
shutting_down_result) into mcp/results.rs, the same R8 LOC-cap move
instructions.rs/schemas.rs already made -- merging PR #1182 pushed mcp.rs
11 lines over the hard cap.
