+++ packages/shared/src/ipc/contracts/notifications.ts
@@ -9,7 +9,9 @@ import type { AppNotification, NotificationOpen, NotificationToast } from '../..
  */
 export interface NotificationsContract {
   list(): Promise<AppNotification[]>;
+  /** Mark one notification read by id — no "mark unread" path exists, so this cannot be undone. */
   markRead(id: string): Promise<void>;
+  /** Mark every notification in the inbox read — no "mark unread" path exists, so this cannot be undone. */
   markAllRead(): Promise<void>;
   /** Permanently delete one notification from the inbox by id. */
   remove(id: string): Promise<void>;
